+++ src/settings.ts
@@ -1,0 +1,4 @@
+export const MODULE_NAME = "shared-token-visibility";
+export enum SettingName {
+    RevealNpc = "revealNpcs",
+}